requirements.txt
# Flask и основные расширения
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Migrate==4.0.4
Flask-SocketIO==5.3.4
Flask-Limiter==3.5.0
Flask-Caching==2.1.0
Flask-CORS==4.0.0
Flask-Compress==1.14
Flask-Talisman==1.1.0
email-validator==2.0.0

# Научные библиотеки и анализ данных
pandas>=1.5.0
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.10.0
seaborn>=0.12.0

# Машинное обучение - основные библиотеки
scikit-learn>=1.3.0
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
tensorflow>=2.13.0

# Техническое анализ и финансы
TA-Lib>=0.4.25
yfinance>=0.2.0
alpha-vantage>=2.3.0

# Криптовалютные API
pycoingecko>=3.2.0
ccxt>=4.0.0

# Обработка данных и утилиты
python-dotenv>=1.0.0
pytz>=2023.3
joblib>=1.3.0
tqdm>=4.65.0

# Веб и развертывание
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
Werkzeug==2.3.7

# Визуализация
plotly>=5.15.0
dash>=2.10.0

# Telegram бот
python-telegram-bot>=20.0

# Кэширование и очереди
redis>=4.5.0
celery>=5.3.0

# Мониторинг и безопасность
sentry-sdk[flask]>=1.32.0
marshmallow>=3.20.0
marshmallow-sqlalchemy>=0.29.0
bleach>=6.0.0
cryptography>=41.0.0
bcrypt>=4.0.0
psutil>=5.9.0

# Дополнительные ML библиотеки для высокой точности
optuna>=3.2.0
hyperopt>=0.2.7
shap>=0.42.0
lime>=0.2.0
imbalanced-learn>=0.11.0
